\name{Illumina450ProbeVariants.db-package}
\alias{Illumina450ProbeVariants.db-package}
\alias{Illumina450ProbeVariants.db}
\alias{450kProbeVariants}
\docType{package}
\title{
Annotation Package combining data from 1000 Genomes Project for Illumina HumanMethylation450 Bead Chip probes
}
\description{
Includes details on variants for each probe on the 450k bead chip for each of the four populations (Asian, American, African and European)
}
\details{
\tabular{ll}{
Package: \tab Illumina450ProbeVariants.db\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 2013-10-07\cr
License: \tab GPL-3\cr
}

}
\author{
Lee Butcher, UCL Cancer Institute, Medical Genomics\cr
Maintainer: Tiffany Morris <tiffany.morris@ucl.ac.uk>

}
\keyword{ package }

